Make RayTracer work again
* Fix naming standard
* Fix some compile warnings due to casting

+++ Raphael/header/UploadBuffer.h
@@ -99,7 +99,7 @@ public:
         return mUploadBuffer.Get();
     }
 
-    void CopyData(int elementIndex, const T& data)
+    void CopyData(size_t elementIndex, const T& data)
     {
         memcpy(&mMappedData[elementIndex * mElementByteSize], &data, sizeof(T));
     }
